fix more in findClosestGraphPositionPathPosition
fix problem with gpp may be in different direction than edge
fix problem with enforcing that only positions ahead in the path are
considered. I was mistakenly treating points closer on the gpp
preferentially. But all i need to do is ensure they are at least ahead
of the reference. Not necessarily close to it.