fix blank screen problem
through some contortions, it is possible to have surfaceChanged called
after onPause, and then onResume is called, but surfaceChanged is not
called. Since surfaceChanged is never handled, the various models are
not rendered and the screen is blank. Previously, I was skipping
handling surfaceChanged when paused, but now I see that sometimes it is
needed.